requirements.txt
azure-kusto-data==0.0.30
azure-kusto-ingest==0.0.30
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
idna==2.8
Markdown==3.1
numpy==1.16.2
pandas==0.24.1
python-dateutil==2.8.0
python-slugify==3.0.0
pytz==2018.9
requests==2.21.0
six==1.12.0
soupsieve==1.9
text-unidecode==1.2
urllib3==1.24.1
